Reduction API

Reduction Recipes

EchelleReduce.ReduceRecipe — Type

ReduceRecipe

Base type for reduction recipes.

Primary reduction methods

EchelleReduce.reduce — Function

reduce(recipe::ReduceRecipe)

Perform reduction steps according to the recipe. The default steps are:

create_output_dirs(recipe)

data = initialize_data(recipe)

gen_master_calib_images(recipe, data)

traces, sregions = get_traces(recipe, data)

extract(recipe, data, traces, sregions)

EchelleReduce.initialize_data — Function

initialize_data

Method intended to initialize parse the input directory and categorize the data. Must be implemented.

EchelleReduce.create_output_dirs — Function

create_output_dirs

Method intended to create any output directories. Must be implemented.

EchelleReduce.gen_master_calib_images — Function

gen_master_calib_images

Method intended to generate all calibration frames. Must be implemented.

EchelleReduce.get_master_flat — Function

get_master_flat

Method intended to get the appropriate master flat frame for calibration. Must be implemented.

EchelleReduce.get_master_dark — Function

get_master_dark

Method intended to get the appropriate master dark frame for calibration. Must be implemented.

EchelleReduce.get_traces — Function

get_traces

Method intended to get the appropriate vector of trace params. Must be implemented.

EchelleReduce.get_trace_spacing — Function

gettracespacing Returns the minimum trace spacing.

EchelleReduce.get_trace_height — Function

gettraceheight Returns the maximum trace height.

EchelleReduce.get_specregion2d — Function

get_specregion2d

Returns the appropriate SpecRegion2d

EchelleReduce.get_extract_orders — Function

get_extract_orders

Gets the orders to be extracted. Must be implemented.

EchelleReduce.get_read_noise — Function

get_read_noise(itime::Real, dark_current::Real=0, read_noise::Real=0)
get_read_noise(data::SpecData, dark_current::Real=0, read_noise::Real=0)

Compute the effective read noise in units of photoelectrons.

EchelleReduce.extract — Function

extract

Method intended to perform all necessary extraction steps. Must be implemented.

Pre-extraction Calibration

EchelleReduce.PreCalib.gen_master_dark — Function

gen_master_dark(master_dark::MasterCal2d; master_bias=nothing)

Generate the master dark image.

EchelleReduce.PreCalib.gen_master_flat — Function

gen_master_flat(master_flat::MasterCal2d; master_bias=nothing, master_dark=nothing, p=0.5)

Generate the master flat field image, normalized to p.

EchelleReduce.PreCalib.pre_calibrate! — Function

pre_calibrate!(image::AbstractMatrix; master_dark=nothing, master_flat=nothing)

Pre-calibrate an image according to the master dark and flat.

EchelleReduce.PreCalib.gen_master_coadded_image — Function

gen_master_coadded_image(master_frame::SpecData2d)

Co-adds frames.

Order Tracing

EchelleReduce.Tracing.trace — Function

trace(data::SpecData2d, sregion::SpecRegion2d; trace_pos_deg=2, min_order_spacing::Int, xleft=nothing, xright=nothing, n_slices=20, fiber=nothing)

Util peaks within a given vector together

data The data

traces The vector of trace parameters (dictionaries)

ny. The number of vertical pixels.

nx. The number of horizontal pixels.

sregion. The spectral region to further bound the left and right ends of the image.

trace(data::SpecData2d, sregion::SpecRegion2d; trace_pos_deg=2, min_order_spacing::Int, xleft=nothing, xright=nothing, n_slices=20, fiber=nothing)

Util peaks within a given vector together

data The data

data_image

traces The vector of trace parameters (dictionaries)

ny. The number of vertical pixels.

nx. The number of horizontal pixels.

sregion. The spectral region to further bound the left and right ends of the image.

EchelleReduce.Tracing.gen_trace_image — Function

gen_trace_image(traces::Vector, ny, nx, sregion::SpecRegion2d)

Generate an nx x ny image with values according to the labels in each trace dictionary.

traces The vector of trace parameters (dictionaries)

ny. The number of vertical pixels.

nx. The number of horizontal pixels.

sregion. The spectral region to further bound the left and right ends of the image.

Spectral Extraction

EchelleReduce.Extract.SpectralExtractor — Type

SpectralExtractor

The base type for a spectral extraction algorithm.

EchelleReduce.Extract.extract_image — Function

extract_image(extractor::SpectralExtractor, data::SpecData2d, data_image::AbstractMatrix, sregion::SpecRegion2d, traces::Vector; badpix_mask::Union{AbstractMatrix, Nothing}=nothing, read_noise::Real=0)

Primary method to extract all orders from an image.

EchelleReduce.Extract.extract_trace — Function

extract_trace

Primary method to extract a single trace. Must be implemented.

EchelleReduce.Extract.gen_model2d — Function

extract_trace

Primary method to generate a 2d model. Must be implemented.

EchelleReduce.Extract.plot_extracted_spectrum — Function

plot_extracted_spectrum(recipe::ReduceRecipe, data::SpecData2d, reduced_data::Vector, fname::String, traces::Vector)

Plots all reduced_data corresponding to traces extracted from a full-frame image and saves the image to fname.

Spectral Extraction Utility Methods

EchelleReduce.Extract.estimate_snr — Function

estimate_snr(trace_image::AbstractMatrix)

Crude method to estimate the mean S/N.

EchelleReduce.Extract.compute_trace_positions_centroids — Function

compute_trace_positions_centroids(trace_image::AbstractMatrix, trace_mask::AbstractMatrix, sregion::SpecRegion2d, trace_positions::Polynomial, aperture; trace_pos_poly_deg=nothing)

Compute the trace positions by fitting a polynomial to the column centroids.

EchelleReduce.Extract.refine_trace_window — Function

refine_trace_window(image::AbstractMatrix, badpix_mask::AbstractMatrix, sregion::SpecRegion2d, trace_positions_estimate::Polynomial; trace_pos_poly_deg=nothing, window::Vector, n_iterations=3)

Iteratively refines the initial trace window using the centroids of each column.

EchelleReduce.Extract.compute_background_1d — Function

compute_background_1d(trace_image::AbstractMatrix, trace_mask::AbstractMatrix, trace_positions::Polynomial, extract_aperture; smooth_width=nothing)

Estimates the background signal by considering regions of low flux according to the provided extract_aperture.

EchelleReduce.Extract.flag_pixels2d! — Function

flag_pixels2d!(trace_image::AbstractMatrix, trace_mask::AbstractMatrix, model2d::AbstractMatrix, nσ::Real)

Flags bad pixels in the 2d image by looking at outliers in the residuals.

EchelleReduce.Extract.fix_bad_pixels_interp — Function

fix_bad_pixels_interp(trace_image::AbstractMatrix, xmin, xmax, poly_bottom::Polynomial, poly_top::Polynomial)

Utility function to fix bad pixels in an image with linear interpolation.

Optimal Spectral Extraction

EchelleReduce.OptimalExtraction.EmpiricalOptimalExtractor — Type

EmpiricalOptimalExtractor(;max_iterations=20, remove_background=false, background_smooth_width=0, oversample_profile=8, trace_pos_poly_deg=4, badpix_σ=6, extract_aperture=:auto)

Construct an EmpiricalOptimalExtractor object with the following parameters.

max_iterations: The maximum number of iterations to run. In each iteration, the trace positions, profile, and bad pixel variables are updated. Convergence is reached when no more bad pixels are found.

remove_background: Whether or not to estimate and remove a background

background_smooth_width: The rolling window size to smooth the background with. Ignored if remove_background is false.

oversample_profile: Integer oversample factor for the trace profile cubic spline.

trace_pos_poly_deg: Polynomial degree for the trace positions, determined from a polynomial fit to the centroids of detector columns.

badpix_σ: Deviations larger than badpix_σ are flagged after extraction.

extract_aperture: The aperture window to consider in extraction. Note that the background will only utilize pixels outside this window. Alternatively, :auto can be passed to automatically determine an optimal window from the profile.

EchelleReduce.OptimalExtraction.optimal_extraction — Function

optimal_extraction(trace_image::AbstractMatrix, trace_mask::AbstractMatrix, trace_positions::Polynomials.Polynomial, trace_profile, extract_aperture, background=nothing, background_err=nothing, read_noise=0, max_iterations=20)

Perform optimal extraction as described in Horne et al. 1985.