Spectral Data API

EchelleBase.SpectralData.SpecData — Type

An abstract type for all spectral data, both 2d echellograms and extracted 1d spectra, parametrized by the spectrograph S.

EchelleBase.SpectralData.SpecData1d — Type

A SpecData1d.

Fields

fname: The starting pixel.

header: The ending pixel.

data: DataFrame.

EchelleBase.SpectralData.SpecData2d — Type

An abstract type for all 2d spectral data (echellograms), parametrized by the spectrograph S.

Missing docstring.

Missing docstring for Echellogram. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RawSpecData2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MasterCal2d. Check Documenter's build log for details.

EchelleBase.SpectralData.get_spectrograph — Function

get_spectrograph(data::SpecData{S})

Returns the name of the spectrograph as a string corresponding to this SpecData object.

EchelleBase.SpectralData.get_spec_module — Function

get_spec_module(data::SpecData{S})

Returns the module for this spectrograph.

Missing docstring.

Missing docstring for read_header. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_image. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_spec1d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ordermin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ordermax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for orderbottom. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ordertop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parse_exposure_start_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parse_itime. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parse_object. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parse_sky_coord. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parse_utdate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_exposure_midpoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_barycentric_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_barycentric_corrections. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_λsolution_estimate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for normalize!. Check Documenter's build log for details.

EchelleBase.SpectralRegions.SpecRegion1d — Type

A container for 1d spectral regions for reduced spectra.

Fields

pixmin: The starting pixel.

pixmax: The ending pixel.

λmin: The starting wavelength.

λmax: The ending wavelength.

order: The echelle order number, if applicable.

label: The label for this chunk.

fiber: The fiber number, if applicable.

EchelleBase.SpectralRegions.SpecRegion2d — Type

A container for 2d spectral regions for echellograms.

Fields

pixmin: The starting pixel in the spectral direction.

pixmax: The ending pixel in the spectral direction.

orderbottom: The bottom order.

ordertop: The top order.

poly_bottom: The bottom bounding polynomial.

poly_top: The top bounding polynomial.

EchelleBase.mask! — Function

mask!(data::SpecData1d, sregion::SpecRegion1d)

Mask bad pixels in-place according to the variables data.data.λ, data.data.flux, data.data.fluxerr, data.data.mask, as well as the bounding pixels in sregion (not wavelength).

mask!(image::AbstractMatrix{<:Number}, sregion::SpecRegion2d)

Mask bad pixels in-place according to the bounding polynomials and left/right ends.