Spectral Modeling API

Barycenter corrections

EchelleSpectralModeling.compute_barycentric_corrections — Function

compute_barycentric_corrections(data::SpecData, star_name=nothing, obs_name=nothing; store=true)
compute_barycentric_corrections(jdmid::Float64, obs_name::String, star_name::String)

Compute the barycentric corrections using barycorrpy for the observation data. The star_name must be a recognized by simbad. If store=true, the variable data.header is updated with the BJD and velocity correction (bjd) and (bc_vel). Returns the BJD and velocity correction as a tuple.

Ensembles - Primary containers

EchelleSpectralModeling.SpectralRVEnsembleProblem — Type

SpectralRVEnsembleProblem{S}

Abstract type for an ensemble of 1d spectral data from a single spectrograph with the primary goal of generating radial velocities.

EchelleSpectralModeling.IterativeSpectralRVEnsembleProblem — Type

Primary container for iteratively generating RVs for a given dataset with a given model and objective function.

Fields

data::Vector{SpecData1d{S}} The vector of SpecData1d objects.

model::AbstractSpectralForwardModel The spectral forward model to use.

obj::SpectralModelObjectiveFunction The objective function to use.

EchelleSpectralModeling.get_init_parameters — Function

get_init_parameters

Get the initial parameters.

EchelleSpectralModeling.load_templates! — Function

load_templates!

Load and store any templates.

Spectral Models

EchelleSpectralModeling.AbstractSpectralForwardModel — Type

AbstractSpectralForwardModel

Abstract type for a spectral forward model.

EchelleSpectralModeling.SpectralForwardModel — Type

SpectralForwardModel{S<:SpectralModelComponent, T<:SpectralModelComponent, G<:SpectralModelComponent, W<:SpectralModelComponent, P<:SpectralModelComponent, C<:SpectralModelComponent}

Primary container for a spectral forward model which models the star, tellurics, the continuum, the wavelength solution, a gas cell (optional), and the line spread function. The wavelength solution may be static using APrioriλSolution.

Fields:

star::SpectralModelComponent` The stellar model.

tellurics::SpectralModelComponent The telluric model. Optional.

gascell::SpectralModelComponent The gas cell model. Optional.

λsolution::SpectralModelComponent The wavelength solution model.

lsf::SpectralModelComponent The line spread function model.

continuum::SpectralModelComponent The continuum model.

sregion::SpecRegion1d The spectral region for this model.

oversample::Int The oversampleing factor of the model.

templates::Dict{String, Array{Float64}} Contains any templates.

Missing docstring.

Missing docstring for load_templates!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_init_parameters. Check Documenter's build log for details.

EchelleSpectralModeling.build — Function

build

Build the model or model component.

Continuum

EchelleSpectralModeling.Continuum.PolyContinuum — Type

PolyContinuum(;deg::Int, coeffs_guess::Vector{Float64})

Construct a PolyContinuum model component of degree deg.

EchelleSpectralModeling.Continuum.SplineContinuum — Type

Container for a SplineContinuum model.

Fields

n_splines::Int The number of splines.

bounds::Vector{Float64}: The bounds for each knot (in units of flux).

Missing docstring.

Missing docstring for estimate_continuum. Check Documenter's build log for details.

Wavelength Solution

EchelleSpectralModeling.Wavelength.APrioriλSolution — Type

APrioriλSolution

An empty type for wavelength solutions known a priori, likely from an LFC, etalon, or ThAr lamp. By default, this grid is stored in the data object as data.data.λ.

EchelleSpectralModeling.Wavelength.SplineλSolution — Type

SplineλSolution(;n_splines::Int, bounds::Vector{Float64})

Construct a SplineλSolution model component with n_splines (# of knots = n_splines + 1). Each knot (in units of wavelength) is bounded by bounds.

EchelleSpectralModeling.Wavelength.PolyλSolution — Type

PolyλSolution(;deg::Int, bounds::Vector{Float64})

Construct a PolyλSolution model component of degree deg. The optimized parameters are set points opposed to coefficients. Each set point is(in units of wavelength) is bounded by bounds.

Combs/Etalons

EchelleSpectralModeling.Wavelength.get_peaks — Function

get_peaks(λ, flux, ν0, Δν; σ_guess=[0.2, 1.4, 3.0], μ_bounds=[-1, 1])

Computes the spot centroids. Returns a tuple:

centers_pixels::Vector{Float64} The pixel centroids.

centers_λ::Vector{Float64} The corresponding wavelength centroids.

integers::Vector{Int64} The corresponding spot integers, relative to ν0.

amplitudes::Vector{Float64} The modeled amplitude of each spot.

σs::Vector{Float64} The modeled sigma of each spot.

rms::Vector{Float64} The RMS from the fit for each spot.

offsets::Vector{Float64} The modeled zero point (in flux, relative to 1) of each spot.

slopes::Vector{Float64} The modeled slope (in flux, relative to 1) of each spot.

Star

EchelleSpectralModeling.Star.AugmentedStar — Type

AugmentedStar(;input_file::String, vel_bounds=[-5000, 5000], star_name::String, absolute_rv_guess::Real=0.0)

Construct an AugmentedStar model component. input_file must be comma delimited. Comments can start with #. The name of the star is used to determine the barycentric corrections with barycorrpy. If input_file is nothing, the model starts from a flat template.

Tellurics

EchelleSpectralModeling.Tellurics.TAPASTellurics — Type

TAPASTellurics(;input_file::String, min_feature_flux=0.02, Δv_guess=[-100, 10, 100], τ_water_guess=[0.1, 1.1, 4], τ_airmass_guess=[0.8, 1.1, 3])

Construct a TAPASTellurics model component. The templates are stored within input_file. C02, N2O, O2, O3, and CH4 are combined into a single template and utilize the parameter airmass_depth. H2O is kept separate and utilizes water_depth. If there are no features less than min_feature_flux, the parameters for this chunk are fixed for the appropriate template.

Line Spread Function (LSF) / Instrument Profile (IP)

EchelleSpectralModeling.LSF.HermiteLSF — Type

HermiteLSF(;deg::Int, σ_guess::Vector{Float64}, coeff_guess::Vector{Float64})

Construct a HermiteLSF model component of degree n, width σ, and coeffs. n=0 corresponds to a pure Gaussian.

Gas Cells

EchelleSpectralModeling.GasCells.GasCell — Type

GasCell(;input_file::String, shift_guess::Vector{Float64}=[0.0, 0.0, 0.0], depth_guess::Vector{Float64}=[1.0, 1.0, 1.0])

Constructs a GasCell model component.

Template Augmenting

EchelleSpectralModeling.TemplateAugmenting.TemplateAugmenter — Type

Base type for a template augmenter.

EchelleSpectralModeling.TemplateAugmenting.WeightedMedianAugmenter — Type

Augment the stellar template variable by computing the weighted median of the residuals in the barycentric frame. Weights are proportional to 1 / rms^2.

EchelleSpectralModeling.TemplateAugmenting.augment_star! — Function

augment_star!(model::SpectralForwardModel, data::Vector{SpecData1d{S}}, opt_results::Vector, augmenter::WeightedMedianAugmenter) where {S}

Augment the stellar template variable by computing the weighted median of the residuals in the barycentric frame. Weights are proportional to 1 / rms^2.

Objective Functions

EchelleSpectralModeling.SpectralModelObjectiveFunctions.Chi2 — Type

Container for a Chi2 objective.

EchelleSpectralModeling.SpectralModelObjectiveFunctions.RMS — Type

Container for a RMS objective (uniform weights).

EchelleSpectralModeling.SpectralModelObjectiveFunctions.compute_obj — Function

compute_obj

Computes the objective function. Must be implemented.