Spectral Data API

EchelleBase.SpectralData.SpecData — Type

An abstract type for all spectral data, both 2d echellograms and extracted 1d spectra, parametrized by the spectrograph symbol S.

EchelleBase.SpectralData.SpecData1d — Type

A SpecData1d.

Fields

fname::String The filename.

header::FITSHeader The fits header.

data::DataFrame A DataFrame containing the actual data.

EchelleBase.SpectralData.SpecData2d — Type

An abstract type for all 2d spectral data (echellograms), parametrized by the spectrograph symbol S.

EchelleBase.SpectralData.Echellogram — Type

Echellogram is an alias for SpecData2d.

EchelleBase.SpectralData.RawSpecData2d — Type

A RawSpecData2d.

Fields

fname::String The filename.

header::FITSHeader The fits header.

EchelleBase.SpectralData.MasterCal2d — Type

A MasterCal2d.

Fields

fname::String The filename.

group::Vector{SpecData2d{S}} The vector of individual SpecData objects used to generate this frame.

EchelleBase.SpectralData.get_spectrograph — Function

SpectralData.get_spectrograph(ensemble::IterativeSpectralRVEnsembleProblem)

Gets the spectrograph as a string for this ensemble.

get_spectrograph(data::SpecData{S})

Returns the name of the spectrograph as a string corresponding to this SpecData object.

EchelleBase.SpectralData.get_spec_module — Function

get_spec_module(data::SpecData{S})

Returns the module for this spectrograph.

EchelleBase.SpectralData.read_header — Function

read_header

Primary method to read in the fits header. Must be implemented.

EchelleBase.SpectralData.read_image — Function

read_image

Primary method to read in an image. Must be implemented.

EchelleBase.SpectralData.read_spec1d — Function

read_spec1d

Primary method to read in a reduced spectrum. Must be implemented.

EchelleBase.SpectralData.ordermin — Function

ordermin

Returns the minimum order on the detector for a given exposure.

EchelleBase.SpectralData.ordermax — Function

ordermax

Returns the maximum order on the detector for a given exposure.

EchelleBase.SpectralData.orderbottom — Function

orderbottom

Returns the bottom order on the detector for a given exposure. It is not necessary that orderbottom < ordertop. Must be implemented.

EchelleBase.SpectralData.ordertop — Function

ordertop

Returns the top order on the detector for a given exposure. Must be implemented.

EchelleBase.SpectralData.parse_exposure_start_time — Function

parse_exposure_start_time

Parses the exposure start time for a given exposure.

EchelleBase.SpectralData.parse_itime — Function

parse_itime

Parses the integration (exposure) time for a given exposure.

EchelleBase.SpectralData.parse_object — Function

parse_object

Parses the object name for a given exposure.

EchelleBase.SpectralData.parse_sky_coord — Function

parse_sky_coord

Parses the sky coordinate for a given exposure.

EchelleBase.SpectralData.parse_utdate — Function

parse_utdate

Parses the UT date for a given exposure.

EchelleBase.SpectralData.get_exposure_midpoint — Function

get_exposure_midpoint

Gets the exposure midpoint.

EchelleBase.SpectralData.get_barycentric_velocity — Function

get_barycentric_velocity

Gets the barycentric velocity correction.

EchelleBase.SpectralData.get_barycentric_corrections — Function

get_barycentric_corrections

Gets the barycentric Julian date and velocity correction.

EchelleBase.SpectralData.get_λsolution_estimate — Function

get_λsolution_estimate

Gets an estimate for the wavelength solution. The precision of the returned grid will be subject to the stability of the spectrograph.

EchelleBase.SpectralData.normalize! — Function

normalize!(data::SpecData1d; p=0.98)

Normalizes the spectral flux and flux error of a 1d spectrum to p.

EchelleBase.SpectralRegions.SpecRegion1d — Type

A container for 1d spectral regions for reduced spectra.

Fields

pixmin: The starting pixel.

pixmax: The ending pixel.

λmin: The starting wavelength.

λmax: The ending wavelength.

order: The echelle order number, if applicable.

label: The label for this chunk.

fiber: The fiber number, if applicable.

EchelleBase.SpectralRegions.SpecRegion2d — Type

A container for 2d spectral regions for echellograms.

Fields

pixmin: The starting pixel in the spectral direction.

pixmax: The ending pixel in the spectral direction.

orderbottom: The bottom order.

ordertop: The top order.

poly_bottom: The bottom bounding polynomial.

poly_top: The top bounding polynomial.

EchelleBase.mask! — Function

mask!(data::SpecData1d, sregion::SpecRegion1d)

Mask bad pixels in-place according to the variables data.data.λ, data.data.flux, data.data.fluxerr, data.data.mask, as well as the bounding pixels in sregion (not wavelength).

mask!(image::AbstractMatrix{<:Number}, sregion::SpecRegion2d)

Mask bad pixels in-place according to the bounding polynomials and left/right ends.