Spectral Modeling API

EchelleSpectralModeling.SpectralRVEnsembleProblem — Type

SpectralRVEnsembleProblem{S}

Abstract type for an ensemble of 1d spectral data from a single spectrograph with the primary goal of generating radial velocities.

EchelleSpectralModeling.IterativeSpectralRVEnsembleProblem — Type

Primary container for iteratively generating RVs for a given dataset with a given model and objective function.

Fields

data::Vector{SpecData1d{S}} The vector of SpecData1d objects.

model::AbstractSpectralForwardModel The spectral forward model to use.

obj::SpectralModelObjectiveFunction The objective function to use.

EchelleSpectralModeling.get_init_parameters — Function

get_init_parameters(m::SpectralModelComponent, args...)

Primary method to get the initial parameters for this model component. Must be implemented.

get_init_parameters(m::SpectralForwardModel, data)

Gets the initial parameters for a given observation.

get_init_parameters(ensemble::IterativeSpectralRVEnsembleProblem)

Gets the initial parameters for all observations.

EchelleSpectralModeling.load_templates! — Function

load_templates!(m::SpectralForwardModel, data)

Loads in the stellar, tellurics, and gas cell templates (if any).

load_templates!(ensemble::IterativeSpectralRVEnsembleProblem)

Loads in the necessary templates for this ensemble.

EchelleSpectralModeling.AbstractSpectralForwardModel — Type

AbstractSpectralForwardModel

Abstract type for a spectral forward model.

EchelleSpectralModeling.SpectralForwardModel — Type

SpectralForwardModel{S<:SpectralModelComponent, T<:SpectralModelComponent, G<:SpectralModelComponent, W<:SpectralModelComponent, P<:SpectralModelComponent, C<:SpectralModelComponent}

Primary container for a spectral forward model which models the star, tellurics, the continuum, the wavelength solution, a gas cell (optional), and the line spread function. The wavelength solution may be static using APrioriλSolution.

Fields:

star::SpectralModelComponent` The stellar model.

tellurics::SpectralModelComponent The telluric model. Optional.

gascell::SpectralModelComponent The gas cell model. Optional.

λsolution::SpectralModelComponent The wavelength solution model.

lsf::SpectralModelComponent The line spread function model.

continuum::SpectralModelComponent The continuum model.

sregion::SpecRegion1d The spectral region for this model.

oversample::Int The oversampleing factor of the model.

templates::Dict{String, Array{Float64}} Contains any templates.

Missing docstring.

Missing docstring for load_templates!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_init_parameters. Check Documenter's build log for details.

EchelleSpectralModeling.build — Function

build(m::SpectralForwardModel, pars::Parameters, data; interp=true)

Builds the model for a given set of parameters and given observation.